vim: skeleton: Use BufNewFile instead of FileType
FileType event may not be publihsed, so Uesr plugin-skeleton-detect also may
not be published.  This behavior is not expected.  So code using FileType is
modified to use BufNewFile instead.

For this change, :setfiletype doesn't cause to set up a skeleton automatically.
But the above problem is worse than this demerit.